TMDM-11887  [Studio] Tab management in the web UI (#321)
* TMDM-13896  [Studio] Update API to support "Category" annotation

* TMDM-13896 Update API to support "Category" annotation.

* TMDM-13896 Fix the following:
1. Fix getFields() can not return fields in base type.
2. Add JUnit test for current "Category" API

* TMDM-13896 Remove blank line.

* TMDM-13896 Fix can not clear category for field and append corresponding JUnit test.

* TMDM-13981 [Studio] Synchronize "Category" annotation after deleting a "field" from entity 

Add test for util methods. make delete of field work for PK field, synchronized annotation for field deletion.

* TMDM-13898  [Studio] Create and Edit "Category" annotation for entity. (#313)

* TMDM-13898 Finish the UI for creating and editing "Category" annotation

* TMDM-13898 Fixed the log output information.

* TMDM-13982 Synchronize "Category" annotation after rename a “field” of entity

* TMDM-13897  Finished displaying "Category" annotation in data model page (#315)

* TMDM-13899  [Studio] Set "Category" annotation for field (#316)

* TMDM-13899 Fixed following:
1. Fix throws NPE when click delete button but no item is selected.
2. Adjust UI layout to avoid text control is stretched when dialog is resized

* TMDM-13899 Fixed delete all field definitions when renaming a category, and append JUnit for it.

* TMDM-14088 Support annotation validation for "category". (#317)

* TMDM-14088 Support annotation validation for "category".

* TMDM-14088 Add JUnit test.

* TMDM-14088 rename var name in messages class and change messages.properties file format from Windows to Unix.

* TMDM-14088 Change file format to Unix

* TMDM-14104 Fixed the following: (#319)

1. Show fields by default order same with defined in entity instead of alphabetical order.
2. Fixed when calling XSDUtil.getFields(), the returned fields include reference field.
3. Add and update JUnit for above fix.

* TMDM-14102 Fixed the following: (#320)

1.Change the label of context menu and title
2. Change all labels which include "field" label to "element".
3. Change the validation message committed in TMDM-14088
4. Add "label" into multi-language dialog
5. Add sort for "Lanuage" and "Label" columns

* TMDM-11887 Fixed still pop "category" context menu when multi select two or more entities or category annotations.

+++ main/plugins/org.talend.mdm.workbench/src/com/amalto/workbench/dialogs/AnnotationLanguageLabelsDialog.java
@@ -14,7 +14,6 @@ package com.amalto.workbench.dialogs;
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Iterator;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.Set;
@@ -23,30 +22,34 @@ import org.eclipse.jface.dialogs.Dialog;
 import org.eclipse.jface.dialogs.IDialogConstants;
 import org.eclipse.jface.dialogs.MessageDialog;
 import org.eclipse.jface.viewers.CellEditor;
+import org.eclipse.jface.viewers.ColumnLabelProvider;
 import org.eclipse.jface.viewers.ComboBoxCellEditor;
 import org.eclipse.jface.viewers.ICellModifier;
-import org.eclipse.jface.viewers.ILabelProviderListener;
+import org.eclipse.jface.viewers.ISelection;
 import org.eclipse.jface.viewers.ISelectionChangedListener;
 import org.eclipse.jface.viewers.IStructuredContentProvider;
 import org.eclipse.jface.viewers.IStructuredSelection;
-import org.eclipse.jface.viewers.ITableLabelProvider;
 import org.eclipse.jface.viewers.SelectionChangedEvent;
 import org.eclipse.jface.viewers.TableViewer;
+import org.eclipse.jface.viewers.TableViewerColumn;
 import org.eclipse.jface.viewers.TextCellEditor;
 import org.eclipse.jface.viewers.Viewer;
+import org.eclipse.jface.viewers.ViewerComparator;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.events.KeyEvent;
 import org.eclipse.swt.events.KeyListener;
 import org.eclipse.swt.events.ModifyEvent;
 import org.eclipse.swt.events.ModifyListener;
+import org.eclipse.swt.events.SelectionAdapter;
+import org.eclipse.swt.events.SelectionEvent;
 import org.eclipse.swt.events.SelectionListener;
-import org.eclipse.swt.graphics.Image;
 import org.eclipse.swt.layout.GridData;
 import org.eclipse.swt.layout.GridLayout;
 import org.eclipse.swt.widgets.Button;
 import org.eclipse.swt.widgets.Combo;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Control;
+import org.eclipse.swt.widgets.Label;
 import org.eclipse.swt.widgets.Shell;
 import org.eclipse.swt.widgets.Table;
 import org.eclipse.swt.widgets.TableColumn;
@@ -72,11 +75,17 @@ public class AnnotationLanguageLabelsDialog extends Dialog {
 
     private String title = "";//$NON-NLS-1$
 
+    private LanguageComparator comparator;
+
+    private static final String LANGUAGE = Messages.AnnotationLanguageLabelsDialog_Language;
+
+    private static final String LABEL = Messages.AnnotationLanguageLabelsDialog_Label;
+
     /**
      * @param parentShell
      */
-    public AnnotationLanguageLabelsDialog(Map<String, String> descriptionsMap, SelectionListener caller,
-            Shell parentShell, String title) {
+    public AnnotationLanguageLabelsDialog(Map<String, String> descriptionsMap, SelectionListener caller, Shell parentShell,
+            String title) {
         super(parentShell);
         if (descriptionsMap != null) {
             this.descriptionsMap = descriptionsMap;
@@ -87,38 +96,44 @@ public class AnnotationLanguageLabelsDialog extends Dialog {
 
     @Override
     protected Control createDialogArea(Composite parent) {
-
         // Should not really be here but well,....
         parent.getShell().setText(this.title);
-
         Composite composite = (Composite) super.createDialogArea(parent);
-
         GridLayout layout = (GridLayout) composite.getLayout();
-        layout.numColumns = 3;
-        // layout.verticalSpacing = 10;
+        layout.numColumns = 4;
+        createDialogContent(composite);
+        return composite;
+    }
 
+    protected void createDialogContent(Composite composite) {
+        Label langLabel = new Label(composite, SWT.NONE);
+        langLabel.setText(Messages.AnnotationLanguageLabelsDialog_LanguageLabel);
+        langLabel.setLayoutData(new GridData(SWT.BEGINNING, SWT.CENTER, false, true, 1, 1));
         languagesCombo = new Combo(composite, SWT.READ_ONLY | SWT.DROP_DOWN | SWT.SINGLE);
         languagesCombo.setLayoutData(new GridData(SWT.BEGINNING, SWT.NONE, false, false, 1, 1));
+
         Set<String> languages = Util.lang2iso.keySet();
-        for (Iterator iter = languages.iterator(); iter.hasNext();) {
-            String language = (String) iter.next();
+        for (Object element : languages) {
+            String language = (String) element;
             languagesCombo.add(language);
         }
         languagesCombo.addModifyListener(new ModifyListener() {
 
+            @Override
             public void modifyText(ModifyEvent e) {
             }
         });
         languagesCombo.select(0);
-
         labelText = new Text(composite, SWT.BORDER | SWT.SINGLE);
-        labelText.setLayoutData(new GridData(SWT.FILL, SWT.FILL, true, true, 1, 1));
+        labelText.setLayoutData(new GridData(SWT.FILL, SWT.NONE, true, false, 1, 1));
         ((GridData) labelText.getLayoutData()).minimumWidth = 150;
         labelText.addKeyListener(new KeyListener() {
 
+            @Override
             public void keyPressed(KeyEvent e) {
             }
 
+            @Override
             public void keyReleased(KeyEvent e) {
                 if ((e.stateMask == 0) && (e.character == SWT.CR)) {
                     String isoCode = Util.lang2iso.get(languagesCombo.getText());
@@ -129,14 +144,16 @@ public class AnnotationLanguageLabelsDialog extends Dialog {
         });
 
         Button addLabelButton = new Button(composite, SWT.PUSH);
-        addLabelButton.setLayoutData(new GridData(SWT.FILL, SWT.FILL, false, true, 1, 1));
+        addLabelButton.setLayoutData(new GridData(SWT.FILL, SWT.FILL, false, false, 1, 1));
         addLabelButton.setImage(ImageCache.getCreatedImage(EImage.ADD_OBJ.getPath()));
         addLabelButton.setToolTipText(Messages.AnnotationLanguageLabelsDialog_Add);
         addLabelButton.addSelectionListener(new SelectionListener() {
 
+            @Override
             public void widgetDefaultSelected(org.eclipse.swt.events.SelectionEvent e) {
             };
 
+            @Override
             public void widgetSelected(org.eclipse.swt.events.SelectionEvent e) {
                 String code = Util.lang2iso.get(languagesCombo.getText());
                 descriptionsMap.put(code, labelText.getText());
@@ -144,38 +161,33 @@ public class AnnotationLanguageLabelsDialog extends Dialog {
             };
         });
 
-        final String LANGUAGE = Messages.AnnotationLanguageLabelsDialog_Language;
-        final String LABEL = Messages.AnnotationLanguageLabelsDialog_Label;
-
-        descriptionsViewer = new TableViewer(composite, SWT.MULTI | SWT.H_SCROLL | SWT.V_SCROLL | SWT.BORDER | SWT.FULL_SELECTION);
-        descriptionsViewer.getControl().setLayoutData(new GridData(SWT.FILL, SWT.FILL, true, true, 2, 1));
+        descriptionsViewer = new TableViewer(composite,
+                SWT.MULTI | SWT.H_SCROLL | SWT.V_SCROLL | SWT.BORDER | SWT.FULL_SELECTION);
+        descriptionsViewer.getControl().setLayoutData(new GridData(SWT.FILL, SWT.FILL, true, true, 3, 1));
         ((GridData) descriptionsViewer.getControl().getLayoutData()).heightHint = 100;
         // Set up the underlying table
         Table table = descriptionsViewer.getTable();
-        // table.setLayoutData(new GridData(GridData.FILL_BOTH));
-
-        new TableColumn(table, SWT.LEFT).setText(LANGUAGE);
-        new TableColumn(table, SWT.CENTER).setText(LABEL);
-        table.getColumn(0).setWidth(150);
-        table.getColumn(1).setWidth(150);
-        for (int i = 2, n = table.getColumnCount(); i < n; i++) {
-            table.getColumn(i).pack();
-        }
-
         table.setHeaderVisible(true);
         table.setLinesVisible(true);
-
+        // Set the sorter for the table
+        comparator = new LanguageComparator();
+        descriptionsViewer.setComparator(comparator);
         Button delLabelButton = new Button(composite, SWT.PUSH);
         delLabelButton.setLayoutData(new GridData(SWT.NONE, SWT.NONE, false, false, 1, 1));
         delLabelButton.setImage(ImageCache.getCreatedImage(EImage.DELETE_OBJ.getPath()));
         delLabelButton.setToolTipText(Messages.AnnotationLanguageLabelsDialog_Del);
         delLabelButton.addSelectionListener(new SelectionListener() {
 
+            @Override
             public void widgetDefaultSelected(org.eclipse.swt.events.SelectionEvent e) {
             };
 
+            @Override
             public void widgetSelected(org.eclipse.swt.events.SelectionEvent e) {
-                deleteItem();
+                ISelection selection = descriptionsViewer.getSelection();
+                if (selection != null && !selection.isEmpty()) {
+                    deleteItem();
+                }
             };
         });
         // Create the cell editors --> We actually discard those later: not natural for an user
@@ -183,23 +195,24 @@ public class AnnotationLanguageLabelsDialog extends Dialog {
         editors[0] = new ComboBoxCellEditor(table, Util.lang2iso.keySet().toArray(new String[] {}), SWT.READ_ONLY);
         editors[1] = new TextCellEditor(table);
         descriptionsViewer.setCellEditors(editors);
-
         // set the content provider
         descriptionsViewer.setContentProvider(new IStructuredContentProvider() {
 
+            @Override
             public void dispose() {
             }
 
+            @Override
             public void inputChanged(Viewer viewer, Object oldInput, Object newInput) {
             }
 
+            @Override
             public Object[] getElements(Object inputElement) {
-                // System.out.println("getElements() ");
                 LinkedHashMap<String, String> descs = (LinkedHashMap<String, String>) inputElement;
                 Set<String> languages = descs.keySet();
                 ArrayList<DescriptionLine> lines = new ArrayList<DescriptionLine>();
-                for (Iterator iter = languages.iterator(); iter.hasNext();) {
-                    String language = ((String) iter.next());
+                for (Object element : languages) {
+                    String language = ((String) element);
                     DescriptionLine line = new DescriptionLine(Util.iso2lang.get(language), descs.get(language));
                     lines.add(line);
                 }
@@ -207,51 +220,18 @@ public class AnnotationLanguageLabelsDialog extends Dialog {
                 return lines.toArray(new DescriptionLine[lines.size()]);
             }
         });
-
-        // set the label provider
-        descriptionsViewer.setLabelProvider(new ITableLabelProvider() {
-
-            public boolean isLabelProperty(Object element, String property) {
-                return false;
-            }
-
-            public void dispose() {
-            }
-
-            public void addListener(ILabelProviderListener listener) {
-            }
-
-            public void removeListener(ILabelProviderListener listener) {
-            }
-
-            public String getColumnText(Object element, int columnIndex) {
-                // System.out.println("getColumnText() "+columnIndex);
-                DescriptionLine line = (DescriptionLine) element;
-                switch (columnIndex) {
-                case 0:
-                    return line.getLanguage();
-                case 1:
-                    return line.getLabel();
-                }
-                return "";//$NON-NLS-1$
-            }
-
-            public Image getColumnImage(Object element, int columnIndex) {
-                return null;
-            }
-        });
-
         // Set the column properties
         descriptionsViewer.setColumnProperties(new String[] { LANGUAGE, LABEL });
-
+        createColumns();
         // set the Cell Modifier
         descriptionsViewer.setCellModifier(new ICellModifier() {
 
+            @Override
             public boolean canModify(Object element, String property) {
-                // if (INSTANCE_ACCESS.equals(property)) return true; Deactivated
                 return true;
             }
 
+            @Override
             public void modify(Object element, String property, Object value) {
                 TableItem item = (TableItem) element;
                 int columnIndex = Arrays.asList(descriptionsViewer.getColumnProperties()).indexOf(property);
@@ -262,7 +242,8 @@ public class AnnotationLanguageLabelsDialog extends Dialog {
                     if (orgIndx != Integer.parseInt(value.toString())) {
                         String newLang = attrs[Integer.parseInt(value.toString())];
                         if (descriptionsMap.containsKey(Util.lang2iso.get(newLang))) {
-                            MessageDialog.openInformation(null, Messages.Warnning, Messages.AnnotationLanguageLabelsDialog_InforContent);
+                            MessageDialog.openInformation(null, Messages.Warnning,
+                                    Messages.AnnotationLanguageLabelsDialog_InforContent);
                             return;
                         }
                         descriptionsMap.remove(Util.lang2iso.get(line.getLanguage()));
@@ -278,6 +259,7 @@ public class AnnotationLanguageLabelsDialog extends Dialog {
                 descriptionsViewer.update(line, null);
             }
 
+            @Override
             public Object getValue(Object element, String property) {
                 int columnIndex = Arrays.asList(descriptionsViewer.getColumnProperties()).indexOf(property);
                 DescriptionLine line = (DescriptionLine) element;
@@ -300,6 +282,7 @@ public class AnnotationLanguageLabelsDialog extends Dialog {
         // Listen for changes in the selection of the viewer to display additional parameters
         descriptionsViewer.addSelectionChangedListener(new ISelectionChangedListener() {
 
+            @Override
             public void selectionChanged(SelectionChangedEvent event) {
             }
         });
@@ -307,12 +290,14 @@ public class AnnotationLanguageLabelsDialog extends Dialog {
         // display for Delete Key events to delete an instance
         descriptionsViewer.getTable().addKeyListener(new KeyListener() {
 
+            @Override
             public void keyPressed(KeyEvent e) {
             }
 
+            @Override
             public void keyReleased(KeyEvent e) {
-                // System.out.println("Table keyReleased() ");
-                if ((e.stateMask == 0) && (e.character == SWT.DEL) && (descriptionsViewer.getSelection() != null)) {
+                ISelection selection = descriptionsViewer.getSelection();
+                if ((e.stateMask == 0) && (e.character == SWT.DEL) && (selection != null && !selection.isEmpty())) {
                     deleteItem();
                 }
             }
@@ -322,8 +307,57 @@ public class AnnotationLanguageLabelsDialog extends Dialog {
         descriptionsViewer.refresh();
 
         labelText.setFocus();
+    }
 
-        return composite;
+    // This will create the columns for the table
+    private void createColumns() {
+        // First column is for the language
+        TableViewerColumn col = createTableViewerColumn(LANGUAGE, 150, 0);
+        col.setLabelProvider(new ColumnLabelProvider() {
+
+            @Override
+            public String getText(Object element) {
+                DescriptionLine line = (DescriptionLine) element;
+                return line.getLanguage();
+            }
+        });
+
+        // Second column is for label
+        col = createTableViewerColumn(LABEL, 150, 1);
+        col.setLabelProvider(new ColumnLabelProvider() {
+
+            @Override
+            public String getText(Object element) {
+                DescriptionLine line = (DescriptionLine) element;
+                return line.getLabel();
+            }
+        });
+    }
+
+    private TableViewerColumn createTableViewerColumn(String title, int bound, final int colNumber) {
+        final TableViewerColumn viewerColumn = new TableViewerColumn(descriptionsViewer, SWT.NONE);
+        final TableColumn column = viewerColumn.getColumn();
+        column.setText(title);
+        column.setWidth(bound);
+        column.setResizable(true);
+        column.setMoveable(true);
+        column.addSelectionListener(getSelectionAdapter(column, colNumber));
+        return viewerColumn;
+    }
+
+    private SelectionAdapter getSelectionAdapter(final TableColumn column, final int index) {
+        SelectionAdapter selectionAdapter = new SelectionAdapter() {
+
+            @Override
+            public void widgetSelected(SelectionEvent e) {
+                comparator.setColumn(index);
+                int dir = comparator.getDirection();
+                descriptionsViewer.getTable().setSortDirection(dir);
+                descriptionsViewer.getTable().setSortColumn(column);
+                descriptionsViewer.refresh();
+            }
+        };
+        return selectionAdapter;
     }
 
     private void deleteItem() {
@@ -404,6 +438,61 @@ public class AnnotationLanguageLabelsDialog extends Dialog {
 
         public void setLanguage(String language) {
             this.language = language;
+        }
+    }
+
+    public class LanguageComparator extends ViewerComparator {
+
+        private int propertyIndex;
+
+        private static final int DESCENDING = 1;
+
+        private int direction = DESCENDING;
+
+        public LanguageComparator() {
+            this.propertyIndex = 0;
+            direction = DESCENDING;
+        }
+
+        public int getDirection() {
+            return direction == 1 ? SWT.DOWN : SWT.UP;
+        }
+
+        public void setColumn(int column) {
+            if (column == this.propertyIndex) {
+                // Same column as last sort; toggle the direction
+                direction = 1 - direction;
+            } else {
+                // New column; do an ascending sort
+                this.propertyIndex = column;
+                direction = DESCENDING;
+            }
+        }
+
+        @Override
+        public int compare(Viewer viewer, Object e1, Object e2) {
+            DescriptionLine p1 = (DescriptionLine) e1;
+            DescriptionLine p2 = (DescriptionLine) e2;
+            int rc = 0;
+            switch (propertyIndex) {
+            case 0:
+                String lang1 = p1.getLanguage() == null ? "" : p1.getLanguage().toLowerCase();
+                String lang2 = p2.getLanguage() == null ? "" : p2.getLanguage().toLowerCase();
+                rc = lang1.compareTo(lang2);
+                break;
+            case 1:
+                String label1 = p1.getLabel() == null ? "" : p1.getLabel().toLowerCase();
+                String label2 = p2.getLabel() == null ? "" : p2.getLabel().toLowerCase();
+                rc = label1.compareTo(label2);
+                break;
+            default:
+                rc = 0;
+            }
+            // If descending order, flip the direction
+            if (direction == DESCENDING) {
+                rc = -rc;
+            }
+            return rc;
         }
     }
 }
